GenericZero > ZeroService
